Add files via upload
This is a Monte Carlo simulation of liquids of two different species mixing at different temperatures. Each molecule of the liquid species for example, oil or water is represented by either 0 or 1s. A ten by ten matrix array filled with 0s and 1s is a visual representation of a container containing two different liquids. The simulation runs and provides a visual plot of how the liquids behave at temperatures in steps of 0.1k between the range of 0.1k to 4.0k. The results of the simulation shows that at low temperatures, immiscible liquids such and oil and water don't mix and are separated. At high temperatures however, the two liquids mix into more homogenous mixtures. This simulation was coded based on physical laws and concepts such as entropy, the Boltzmann distribution and energies of the molecules.